Minor upates on line 117, 224, and 398 to row conversions from i64 to i32

+++ src/db.rs
@@ -174,7 +174,7 @@ pub async fn insert_policy(policy: &policy::Policy) -> Result<u64> {
             .bind(allowed_build_ids_json)
             .fetch_one(&dbpool)
             .await?;
-        Ok(last_insert_row.try_get::<i64, _>(0)? as u64)
+        Ok(last_insert_row.try_get::<i32, _>(0)? as u64)
     }
 }
 
@@ -221,7 +221,7 @@ pub async fn get_secret_policy(sec: &str) -> Result<policy::Policy> {
         .bind(sec)
         .fetch_one(&dbpool)
         .await?;
-    let pol = pol_row.try_get::<i64, _>(0)? as u64;
+    let pol = pol_row.try_get::<i32, _>(0)? as u64;
     let secret_policy = get_policy(pol).await?;
     Ok(secret_policy)
 }
@@ -395,7 +395,7 @@ pub async fn get_signing_keys_policy(key_id: &str) -> Result<Option<policy::Poli
         .await?;
     match policy_id_option {
         Some(p) => {
-            let pid = p.try_get::<i64, _>(0)? as u64;
+            let pid = p.try_get::<i32, _>(0)? as u64;
             Ok(Some(get_policy(pid).await?))
         }
         None => Ok(None),
